RTS, RTR: Fixed reported PC on address and bus error
The JSON based tests I was using showed that when "returned" address of
RTS and RTR points to invalid address, the PC should point to next
instruction of the return instruction, instead of the new instruction
address.
So this commit adds special handling for such case. Right now this is
applied to both address and bus error, but I don't know if it actually
applies to the bus error as well. Well, we'll see how it goes.

Note that I haven't applied the fix to the RTE yet, because it seems
like RTE has some other issues that I need to fix first.